물레방아, 샷건 아이템 구현

+++ NGP_project/Client/TimeManager.cpp
@@ -73,7 +73,7 @@ void TimeManager::Tick(float lockFps)
 	if (_sampleCount > 0) _deltaTime /= _sampleCount;
 }
 
-bool TimeManager::CheckTimer(int& currentTime, int targetTime)
+bool TimeManager::CheckTimer(float& currentTime, float targetTime)
 {
 	currentTime += GetDeltaTime();
